Website/requirements.txt
app==0.0.1
category-encoders==2.2.2
click==7.1.2
cycler==0.10.0
dill==0.3.3
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.3
pandas==1.1.4
patsy==0.5.1
Pillow==8.0.1
pyfunctional==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn-utils==0.0.15
statsmodels==0.12.1
tabulate==0.8.7
threadpoolctl==2.1.0
Werkzeug==1.0.1
xgboost
